채력재생, 레벨업 추가, 보스 추가

+++ Assets/Scripts/Enemy/EnemyProjectile.cs
@@ -5,7 +5,7 @@ using UnityEngine;
 public class EnemyProjectile : MonoBehaviour
 {
     [SerializeField]
-    private int damage = 1;
+    private float damage = 1;
     private void OnEnable()
     {
         gameObject.GetComponent<Movement>().Move(Vector3.up);
